Add local file system monitoring without VFS Create a new module for local file monitoring based on JAVA WatchService API. Move remote file monitoring which based on VFS to a separate module.
