--- file/org.wso2.carbon.transport.file-system/src/main/java/org/wso2/carbon/transport/filesystem/connector/server/contract/FileSystemMessage.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2017 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.transport.filesystem.connector.server.contract;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.wso2.carbon.messaging.CarbonMessage;
-
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-
-/**
- * This class represent the file message.
- */
-public class FileSystemMessage extends CarbonMessage {
-
-    private static final Logger LOG = LoggerFactory.getLogger(FileSystemMessage.class);
-    private final String text;
-
-    public FileSystemMessage(String text) {
-        this.text = text;
-    }
-
-    public String getText() {
-        return this.text;
-    }
-
-    public InputStream getInputStream() {
-        return this.text == null ? null : new ByteArrayInputStream(this.text.getBytes(StandardCharsets.UTF_8));
-    }
-
-    public ByteBuffer getMessageBody() {
-        this.setEndOfMsgAdded(true);
-        return ByteBuffer.wrap(this.text.getBytes(StandardCharsets.UTF_8));
-    }
-}
