--- file/org.wso2.carbon.transport.file-system/src/main/java/org/wso2/carbon/transport/filesystem/connector/server/contract/FileSystemServerConnector.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2017 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.transport.filesystem.connector.server.contract;
-
-import org.wso2.carbon.transport.filesystem.connector.server.exception.FileSystemServerConnectorException;
-
-/**
- * File System Server Connector interface to start and stop the pooling operation for given the directory location.
- */
-public interface FileSystemServerConnector {
-
-    /**
-     * Start the polling operation.
-     *
-     * @throws FileSystemServerConnectorException If unable to start the polling.
-     */
-    void start() throws FileSystemServerConnectorException;
-
-    /**
-     * Stop the polling operation.
-     *
-     * @throws FileSystemServerConnectorException If unable to stop the polling.
-     */
-    void stop() throws FileSystemServerConnectorException;
-}
