+++ file/org.wso2.carbon.transport.remote-file-system/src/main/java/org/wso2/carbon/transport/remotefilesystem/server/connector/util/ThreadPoolFactory.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2017 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.transport.remotefilesystem.server.connector.util;
+
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+
+/**
+ * A Thread Pool factory for creating Thread Pools required for the File System transport
+ */
+public class ThreadPoolFactory {
+    private static ThreadPoolFactory instance;
+    private ExecutorService executorService;
+
+    private ThreadPoolFactory(int threadPoolSize, boolean parallel) {
+        if (parallel) {
+            executorService = Executors.newFixedThreadPool(threadPoolSize);
+        } else {
+            executorService = Executors.newSingleThreadExecutor();
+        }
+    }
+
+    public static synchronized void createInstance(int threadPoolSize, boolean parallel) {
+        if (instance == null) {
+            instance = new ThreadPoolFactory(threadPoolSize, parallel);
+        }
+    }
+
+    public static ThreadPoolFactory getInstance() {
+        return instance;
+    }
+
+    public ExecutorService getExecutor() {
+        return executorService;
+    }
+}
