+++ file/org.wso2.carbon.transport.remote-file-system/src/test/resources/fs-cron/testCron.txt
@@ -1,0 +1,1 @@
+File Server Connector cron test
