+++ file/org.wso2.carbon.transport.remote-file-system/src/test/resources/test/test.txt
@@ -1,0 +1,1 @@
+File Server Connector test
