+++ file/pom.xml
@@ -27,6 +27,7 @@
     <name>WSO2 Carbon Transport File Parent</name>
     <modules>
         <module>org.wso2.carbon.transport.file</module>
-        <module>org.wso2.carbon.transport.file-system</module>
+        <module>org.wso2.carbon.transport.remote-file-system</module>
+        <module>org.wso2.carbon.transport.local-file-system</module>
     </modules>
 </project>
